+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Reformatted section titles in Testing user guides
 - Added supression for RestSharp 'Parameter is Obsolete' warnings
 - Added ICacheableQuestion interface implementation to RestApiDownload (Equals and GetHashCode)
+- Added unit tests for Question that derive from `AbstractWebQuestion`
 
 
 ### Changed